requirements.txt
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.10.0
docopt==0.6.2
idna==2.10
Js2Py==0.71
kiwisolver==1.3.1
matplotlib==3.4.2
MouseInfo==0.1.3
numpy==1.20.3
opencv-python==4.5.2.52
packaging==20.9
Pillow==8.2.0
pipwin==0.5.1
PyAudio @ file:///C:/Users/so4re/pipwin/PyAudio-0.2.11-cp39-cp39-win_amd64.whl
PyAutoGUI==0.9.52
pycparser==2.20
pydub==0.25.1
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==1.0.9
pyparsing==2.4.7
pyperclip==1.8.2
PyPrind==2.11.3
PyRect==0.1.4
PyScreeze==0.1.27
pySmartDL==1.3.4
python-dateutil==2.8.1
PyTweening==1.0.3
pytz==2021.1
requests==2.25.1
selenium==3.141.0
six==1.16.0
SoundFile==0.10.3.post1
soupsieve==2.2.1
tzlocal==2.1
urllib3==1.26.4
